Code refactoring
Move the SSH functions to helpers.js.
