+++ helpers.js
@@ -1,4 +1,5 @@
 const { access, lstat, readFile } = require("fs/promises");
+const { Client } = require("ssh2");
 
 /**
  * Common errors messages
@@ -112,6 +113,35 @@ async function promiseQueue(promiseInitiators) {
   return results;
 }
 
+/**
+ * Creates SSH Client
+ * @param {import("ssh2").ConnectConfig} connectConfig
+ * @returns {Client}
+ */
+function createSSHConnection(connectConfig) {
+  const sshClient = new Client();
+  return new Promise((res, rej) => { sshClient.connect(connectConfig).on("ready", () => res(sshClient)).on("error", rej); });
+}
+
+/**
+ * Executes commands over SSH Client
+ * @param {Client} sshClient
+ * @param {string} cmd
+ * @param {{ endConnectionAfter: boolean }} options
+ */
+function executeOverSSH(sshClient, cmd, { endConnectionAfter = true } = {}) {
+  return new Promise((res, rej) => {
+    sshClient.exec(cmd, (error, channel) => {
+      if (error) return rej(error);
+      if (endConnectionAfter) channel.on("close", () => sshClient.end());
+
+      // handleChildProcess can be used here because the SSH Stream has the same
+      // methods and the same events as the child process
+      return handleChildProcess(channel, { finishSignal: "close" }).then(res).catch(rej);
+    });
+  });
+}
+
 module.exports = {
   joinCommand,
   pathExists,
@@ -120,5 +150,7 @@ module.exports = {
   handleCommonErrors,
   promiseQueue,
   readKeyFile,
+  createSSHConnection,
+  executeOverSSH,
   ERROR_MESSAGES,
 };
